Issue #18862: Enable select mode from main tray menu
